Move LeafPartitionRecordRetriever to query-core
