+++ java/query/query-runner/src/main/java/sleeper/query/runner/recordretrieval/LeafPartitionQueryExecutor.java
@@ -29,6 +29,7 @@ import sleeper.core.properties.table.TableProperty;
 import sleeper.core.record.Record;
 import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
+import sleeper.query.core.LeafPartitionRecordRetriever;
 import sleeper.query.core.RecordRetrievalException;
 import sleeper.query.core.model.LeafPartitionQuery;
 import sleeper.query.core.model.QueryException;
