+++ java/query/query-runner/src/test/java/sleeper/query/runner/recordretrieval/InMemoryDataStore.java
@@ -21,6 +21,7 @@ import sleeper.core.key.Key;
 import sleeper.core.range.Region;
 import sleeper.core.record.Record;
 import sleeper.core.schema.Schema;
+import sleeper.query.core.LeafPartitionRecordRetriever;
 import sleeper.query.core.RecordRetrievalException;
 import sleeper.query.core.model.LeafPartitionQuery;
